pyproject.toml
[tool.poetry]
name = "simples-crud"
version = "0.1.0"
description = ""
authors = ["Joao Pedro Watson <joaopedrowatson@hotmail.com>"]
readme = "README.md"
packages = [{include = "simples_crud"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-cors = "^4.0.0"
flask-migrate = "^4.0.4"
flask-marshmallow = "^0.15.0"
marshmallow-sqlalchemy = "^0.29.0"
alchemical = "^0.7.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.12.0"
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
taskipy = "^1.11.0"


[tool.poetry.group.doc.dependencies]
apifairy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
run = "flask run"
lint = "blue --check --diff . && isort --check --diff ."
auto-lint = "blue . && isort ."
pre_test = "task lint"
test = "pytest -s -x --cov=notas_musicais -vv"
post_test = "coverage html"